requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Vector Database
qdrant-client>=1.7.0

# PostgreSQL Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP Client for external APIs
httpx>=0.25.0
aiohttp>=3.8.0

# Text Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Web Interface
jinja2>=3.1.0
